GafferSceneUI : move OSL texture evaluation from GafferArnoldUI to here
Note: this means we need to link GafferSceneUI to GafferOSL
