+++ src/GafferSceneUI/StandardLightVisualiser.cpp
@@ -34,6 +34,8 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
+#include "GafferOSL/ShadingEngine.h"
+
 #include "GafferSceneUI/StandardLightVisualiser.h"
 
 #include "Gaffer/Metadata.h"
@@ -49,6 +51,8 @@
 
 #include "IECoreScene/MeshPrimitive.h"
 #include "IECoreScene/Shader.h"
+
+#include "IECore/LRUCache.h"
 
 using namespace std;
 using namespace boost;
@@ -208,6 +212,124 @@ const char *environmentSphereFragSource()
 	;
 }
 
+// Functionality for efficiently evaluating an OSL shader network on a grid of points
+
+struct OSLTextureCacheGetterKey
+{
+
+	OSLTextureCacheGetterKey()
+		:	shaderNetwork( nullptr ), resolution( 512 )
+	{
+	}
+
+	OSLTextureCacheGetterKey( const IECoreScene::ShaderNetwork *shaderNetwork, int resolution )
+		:	shaderNetwork( shaderNetwork ), resolution( resolution )
+	{
+		shaderNetwork->hash( hash );
+		hash.append( resolution );
+	}
+
+	operator const IECore::MurmurHash & () const
+	{
+		return hash;
+	}
+
+	const IECoreScene::ShaderNetwork *shaderNetwork;
+	int resolution;
+	MurmurHash hash;
+
+};
+
+CompoundDataPtr evalOSL( const IECoreScene::ShaderNetwork *shaderNetwork, int resolution )
+{
+	GafferOSL::ShadingEnginePtr shadingEngine = new GafferOSL::ShadingEngine( shaderNetwork );
+
+	CompoundDataPtr shadingPoints = new CompoundData();
+
+	V3fVectorDataPtr pData = new V3fVectorData;
+	FloatVectorDataPtr uData = new FloatVectorData;
+	FloatVectorDataPtr vData = new FloatVectorData;
+
+	vector<V3f> &pWritable = pData->writable();
+	vector<float> &uWritable = uData->writable();
+	vector<float> &vWritable = vData->writable();
+
+	int numPoints = resolution * resolution;
+
+	pWritable.reserve( numPoints );
+	uWritable.reserve( numPoints );
+	vWritable.reserve( numPoints );
+
+	for( int y = 0; y < resolution; ++y )
+	{
+		for( int x = 0; x < resolution; ++x )
+		{
+			uWritable.push_back( (float)(x + 0.5f) / resolution );
+			// V is flipped because we're generating a Cortex image,
+			// and Cortex has the pixel origin at the top left.
+			vWritable.push_back( 1.0f - ( (y + 0.5f) / resolution ) );
+			pWritable.push_back( V3f( x + 0.5f, y + 0.5f, 0.0f ) );
+		}
+	}
+
+	shadingPoints->writable()["P"] = pData;
+	shadingPoints->writable()["u"] = uData;
+	shadingPoints->writable()["v"] = vData;
+
+	CompoundDataPtr shadingResult = shadingEngine->shade( shadingPoints.get() );
+	ConstColor3fVectorDataPtr colors = shadingResult->member<Color3fVectorData>( "Ci" );
+
+	CompoundDataPtr result = new CompoundData();
+
+	if( colors )
+	{
+		Imath::Box2i dataWindow( Imath::V2i( 0.0f ), Imath::V2i( resolution - 1 ) );
+		Imath::Box2i displayWindow( Imath::V2i( 0.0f ), Imath::V2i( resolution - 1 ) );
+
+		result->writable()["dataWindow"] = new Box2iData( dataWindow );
+		result->writable()["displayWindow"] = new Box2iData( displayWindow );
+
+		FloatVectorDataPtr redChannelData = new FloatVectorData();
+		FloatVectorDataPtr greenChannelData = new FloatVectorData();
+		FloatVectorDataPtr blueChannelData = new FloatVectorData();
+		std::vector<float> &r = redChannelData->writable();
+		std::vector<float> &g = greenChannelData->writable();
+		std::vector<float> &b = blueChannelData->writable();
+
+		vector<Color3f>::size_type numColors = colors->readable().size();
+		r.reserve( numColors );
+		g.reserve( numColors );
+		b.reserve( numColors );
+
+		for( vector<Color3f>::size_type u = 0; u < numColors; ++u )
+		{
+			Color3f c = colors->readable()[u];
+
+			r.push_back( c[0] );
+			g.push_back( c[1] );
+			b.push_back( c[2] );
+		}
+
+		CompoundDataPtr channelData = new CompoundData;
+		channelData->writable()["R"] = redChannelData;
+		channelData->writable()["G"] = greenChannelData;
+		channelData->writable()["B"] = blueChannelData;
+
+		result->writable()["channels"] = channelData;
+	}
+
+	return result;
+}
+
+CompoundDataPtr getter( const OSLTextureCacheGetterKey &key, size_t &cost )
+{
+	cost = 1;
+	return evalOSL( key.shaderNetwork, key.resolution );
+}
+
+typedef LRUCache<IECore::MurmurHash, CompoundDataPtr, LRUCachePolicy::Parallel, OSLTextureCacheGetterKey> OSLTextureCache;
+OSLTextureCache g_oslTextureCache( getter, 100 );
+
 } // namespace
 
 //////////////////////////////////////////////////////////////////////////
@@ -355,6 +477,11 @@ void StandardLightVisualiser::spotlightParameters( const InternedString &attribu
 	{
 		lensRadius = parameter<float>( metadataTarget, shaderParameters, "lensRadiusParameter", 0.0f );
 	}
+}
+
+IECore::CompoundDataPtr StandardLightVisualiser::computeTextureFromOSLNetwork( const IECoreScene::ShaderNetwork *shaderNetwork, int resolution )
+{
+	return g_oslTextureCache.get( OSLTextureCacheGetterKey( shaderNetwork, resolution ) );
 }
 
 const char *StandardLightVisualiser::faceCameraVertexSource()
